Extract BLE loop into bluetooth_loop_linux.go and remove noBT
Move bleLoop, decideBleAction, reconcileBluetooth, bluetoothDesired,
and backoff helpers into a dedicated bluetooth_loop_linux.go file for
easier review. The low-level BlueZ primitives stay in bluetooth_linux.go.

Delete the noBT sticky-failure flag from the Subsystem struct. A single
BlueZ failure previously disabled BLE for the entire process lifetime.
Bounded exponential backoff (5s doubling to 10m) replaces it: transient
failures self-heal on the next successful attempt, and once backoff
reaches the 10m cap retries stop entirely. A log line at that point
suggests setting disable_bt_provisioning if the device lacks bluetooth.

Simplify bluetoothEnabled to just check DisableBTProvisioning config.
